+++ nanos-lite/src/mm.c
@@ -23,7 +23,7 @@ int mm_brk(uintptr_t brk) {
 
 void init_mm() {
   pf = (void *)ROUNDUP(heap.start, PGSIZE);
-  Log("free physical pages starting from 0x%08x", (uint32_t)pf);
+  Log("free physical pages starting from 0x%08x", (uintptr_t)pf);
 
 #ifdef HAS_VME
   vme_init(pg_alloc, free_page);
