+++ nanos-lite/src/ramdisk.c
@@ -25,7 +25,7 @@ size_t ramdisk_write(const void *buf, size_t offset, size_t len) {
 
 void init_ramdisk() {
   Log("ramdisk info: start = 0x%08x, end = 0x%08x, size = 0x%08x bytes",
-      (uint32_t)&ramdisk_start, (uint32_t)&ramdisk_end, RAMDISK_SIZE);
+      (uintptr_t)&ramdisk_start, (uintptr_t)&ramdisk_end, RAMDISK_SIZE);
 }
 
 size_t get_ramdisk_size() {
